+++ richat/src/grpc/server.rs
@@ -47,7 +47,7 @@ use {
         time::{Duration, SystemTime},
     },
     tonic::{
-        service::interceptor::interceptor, Request, Response, Result as TonicResult, Status,
+        service::interceptor::InterceptorLayer, Request, Response, Result as TonicResult, Status,
         Streaming,
     },
     tracing::{error, info, warn},
@@ -148,7 +148,7 @@ impl GrpcServer {
         // Spawn server
         let server = tokio::spawn(async move {
             if let Err(error) = server_builder
-                .layer(interceptor(move |request: Request<()>| {
+                .layer(InterceptorLayer::new(move |request: Request<()>| {
                     if config.x_token.is_empty() {
                         Ok(request)
                     } else {