+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Features
 
 - cli: add support of SubscribeReplayInfo ([#133](https://github.com/lamports-dev/richat/pull/133))
+- richat: update yellowstone-grpc-proto ([#134](https://github.com/lamports-dev/richat/pull/134))
 
 ### Breaking